src/backend/requirements.txt
flask>=3.0.0
strawberry-graphql[flask]>=0.220.0
flask-cors>=4.0.0
PyJWT>=2.8
gunicorn
